nms-10306: Add docs, move Enum to common daemon
