+++ opennms-doc/guide-admin/src/asciidoc/index.adoc
@@ -279,6 +279,8 @@ include::text/operation/daemon-config-files/introduction.adoc[]
 include::text/operation/daemon-config-files/eventd.adoc[]
 include::text/operation/daemon-config-files/notifd.adoc[]
 include::text/operation/daemon-config-files/pollerd.adoc[]
+include::text/operation/daemon-config-files/syslogd.adoc[]
+include::text/operation/daemon-config-files/trapd.adoc[]
 
 [[ga-opennms-system-properties]]
 == System Properties
