+++ opennms-doc/guide-admin/src/asciidoc/text/operation/daemon-config-files/syslogd.adoc
@@ -1,0 +1,23 @@
+
+// Allow GitHub image rendering
+:imagesdir: ../../../images
+
+[[ga-opennms-operation-daemon-config-files-syslogd]]
+==== Syslogd
+
+[options="header, autowidth"]
+|===
+| Internal Daemon Name | Reload Event
+| _Syslogd_            | `uei.opennms.org/internal/reloadDaemonConfig -p 'daemonName Syslogd'`
+|===
+
+Syslogd reload event stops and starts daemon and loads all the syslogd configuration changes.
+
+Syslog daemon can be reloaded with following shell command on karaf.
+
+[source]
+----
+$ ssh -p 8101 admin@localhost
+...
+admin@opennms> reload:daemon Syslogd
+----
